Refactor chemistry commands to the tardischemistry package
